chore(gc): split barrier.rs at the 2000-line cap — maintenance helpers to a sibling module
barrier.rs reached 2,157 lines after the 2026-08-11 evening merges. The
remembered-set inspection/drain/clear group moves to barrier/maintenance.rs
(the #7830 recipe: cohesive group, explicit re-export, no logic change);
tail visibilities widen one module level (pub(super) -> pub(in crate::gc))
because the re-export cannot widen them.
